requirements.txt
azure-cognitiveservices-speech>=1.34.0
pyaudio>=0.2.13
numpy>=1.24.0
pydub>=0.25.1
audioop-lts>=0.2.0
colorama>=0.4.6
rich>=13.7.0
pyyaml>=6.0.1
openai>=1.0.0
argparse>=1.4.0